tests/wpt/tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.svg.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.drawImage.svg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.svg</h1>
<p class="desc">drawImage() of an SVG image</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  const img = new Image();
  const imageLoadPromise = new Promise((resolve, reject) => {
    img.onload = () => resolve();
    img.onerror = (err) => reject(err);
  });
  img.src = '/images/green.svg';
  await imageLoadPromise;

  ctx.drawImage(img, 0, 0);
  _assertPixelApprox(canvas, 50, 25, 0, 255, 0, 255, 2);

}, "drawImage() of an SVG image");
</script>
